+++ applications/search/src/components/filter-box/filter-box.component.jsx
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import { Collapse } from 'reactstrap';
 import cx from 'classnames';
 
-import localization from '../../../../lib/localization';
-import { FilterOption } from '../../filter-option/filter-option.component';
+import localization from '../../lib/localization';
+import { FilterOption } from '../filter-option/filter-option.component';
 import './filter-box.scss';
 
 export class FilterBox extends React.Component {
