+++ applications/search/src/components/filter-box/filter-box.scss
@@ -1,4 +1,4 @@
-@import '../../../../../node_modules/designsystemet/fdk-designsystem-bootstrap4/scss/colors.scss';
+@import '../../../node_modules/designsystemet/fdk-designsystem-bootstrap4/scss/colors.scss';
 
 .fdk-panel--filter, .search-filter-publisher {
   background-color: $color5;
