+++ applications/search/src/components/filter-option/filter-option.component.test.jsx
